code-cleanup & logging
