+++ resources/login.py
@@ -14,8 +14,7 @@ def login():
     login_progress = xbmcgui.DialogProgress()
     login_progress.create('Netflix', utility.get_string(30200) + '...')
     utility.progress_window(login_progress, 25, utility.get_string(30201))
-    response = connect.load_netflix_site(utility.main_url + 'Login', new_session=True)
-    content = utility.decode(response)
+    content = connect.load_netflix_site(utility.main_url + 'Login', new_session=True)
     if not 'Sorry, Netflix ' in content:
         match = re.compile('name="authURL" value="(.+?)"', re.DOTALL).findall(content)
 #        utility.log('Setting authorization url: ' + match[0])
@@ -25,9 +24,8 @@ def login():
         post_data = {'authURL': utility.get_setting('authorization_url'), 'email': utility.get_setting('username'),
                      'password': utility.get_setting('password'), 'RememberMe': 'on'}
         utility.progress_window(login_progress, 50, utility.get_string(30202))
-        response = connect.load_netflix_site(utility.main_url + 'Login?locale=' + utility.get_setting('language'),
+        content = connect.load_netflix_site(utility.main_url + 'Login?locale=' + utility.get_setting('language'),
                                          post=post_data)
-        content = utility.decode(response)
 #        utility.log(response)
 
         if 'id="page-LOGIN"' in content:
@@ -36,15 +34,6 @@ def login():
         match = re.compile('"apiUrl":"(.+?)",').findall(content)
         utility.set_setting('api_url', match[0])
         connect.set_chrome_netflix_cookies()
-#        post_data = utility.my_list % utility.get_setting('authorization_url')
-#        content = utility.decode(connect.load_netflix_site(utility.evaluator(), post=post_data))
-#        matches = json.loads(content)['value']
-#        match = matches['lolomos'].keys()
-#        utility.set_setting('root_list', match[0])
-#        match = matches['lists'].keys()
-#        utility.set_setting('my_list', match[1])
-#        match = matches['lists'][utility.get_setting('my_list')]['trackIds']['trackId']
-#        utility.set_setting('track_id', unicode(match))
         utility.progress_window(login_progress, 75, utility.get_string(30203))
         if not (utility.get_setting('selected_profile') or (utility.get_setting('single_profile') == 'true')):
             profiles.choose()
