+++ resources/profiles.py
@@ -27,7 +27,7 @@ def switch_profile(profile_id):
 
 def choose():
     profiles = []
-    content = utility.decode(connect.load_netflix_site(utility.profile_url))
+    content = connect.load_netflix_site(utility.profile_url)
     match = json.loads(content)['profiles']
     for item in match:
         profile = {'name': item['firstName'], 'token': item['guid'], 'is_kid': item['experience'] == 'jfk'}
